refactor(#151): dedup the two WASM adapters into adapter_common
FetchDataSource and LocalSqlDataSource shared ~150 byte-identical lines.
Hoisted the five self-free helpers (rows_to_maps, build_pk_text_expr,
row_maps_to_metadata, content_hash, generate_batch_sql) into a new
adapter_common module as pub(crate) free fns; both adapters now call
adapter_common::<fn>. Pure move -- hashes and SQL strings are byte-identical
to before. cached_table_info stays per-adapter (reads adapter state).

Net ~-145 LOC across the touched files. Also fixes the 10 pre-existing
needless_borrow lints in lib.rs (the lint half of #155) so the wasm-target
clippy -D warnings gate this issue requires passes.

Scope note: the cross-crate pk_text_expr divergence (core's bare "col" vs
plugin/WASM's CAST(... AS TEXT)) is split out to #166 as a correctness fix;
this PR does NOT change pk_text_expr behavior.

wasm build + wasm clippy + host clippy -D warnings + fmt clean.
