+++ crates/smugglr-wasm/src/local_adapter.rs
@@ -6,11 +6,11 @@
 //! better-sqlite3 (Node), official sqlite-wasm, or sql.js without
 //! changes here.
 
-use sha2::{Digest, Sha256};
-use smugglr_core::config::column_excluded;
 use smugglr_core::datasource::{ColumnInfo, DataSource, RowMeta, TableInfo};
 use smugglr_core::error::{Result, SyncError};
 use std::collections::HashMap;
+
+use crate::adapter_common;
 
 use serde_json::Value;
 use wasm_bindgen::prelude::*;
@@ -68,90 +68,6 @@ impl LocalSqlDataSource {
             .map_err(|e| SyncError::Remote(format!("invalid executor result shape: {}", e)))
     }
 
-    fn rows_to_maps(&self, columns: &[String], rows: &[Vec<Value>]) -> Vec<HashMap<String, Value>> {
-        rows.iter()
-            .map(|row| {
-                columns
-                    .iter()
-                    .zip(row.iter())
-                    .map(|(col, val)| (col.clone(), val.clone()))
-                    .collect()
-            })
-            .collect()
-    }
-
-    fn build_pk_text_expr(primary_key: &[String]) -> String {
-        if primary_key.len() == 1 {
-            format!("CAST(\"{}\" AS TEXT)", primary_key[0])
-        } else {
-            primary_key
-                .iter()
-                .map(|k| format!("CAST(\"{}\" AS TEXT)", k))
-                .collect::<Vec<_>>()
-                .join(" || '|' || ")
-        }
-    }
-
-    fn row_maps_to_metadata(
-        maps: &[HashMap<String, Value>],
-        column_order: &[String],
-        timestamp_column: &str,
-        exclude_columns: &[String],
-    ) -> HashMap<String, RowMeta> {
-        let mut result = HashMap::with_capacity(maps.len());
-        for row in maps {
-            let pk = row
-                .get("__pk")
-                .and_then(|v| v.as_str())
-                .unwrap_or("")
-                .to_string();
-            let updated_at = row
-                .get(timestamp_column)
-                .and_then(|v| v.as_str())
-                .map(String::from);
-            let content_hash =
-                Self::content_hash(row, column_order, exclude_columns, timestamp_column);
-            result.insert(
-                pk.clone(),
-                RowMeta {
-                    pk_value: pk,
-                    updated_at,
-                    content_hash,
-                },
-            );
-        }
-        result
-    }
-
-    fn content_hash(
-        row: &HashMap<String, Value>,
-        columns_in_order: &[String],
-        exclude: &[String],
-        timestamp_column: &str,
-    ) -> String {
-        let timestamp_columns = ["updated_at", "created_at"];
-        let mut hasher = Sha256::new();
-        for col in columns_in_order {
-            if timestamp_columns.contains(&col.as_str())
-                || column_excluded(col, exclude)
-                || col == timestamp_column
-            {
-                continue;
-            }
-            if let Some(val) = row.get(col) {
-                match val {
-                    Value::Null => {}
-                    Value::String(s) => hasher.update(s.as_bytes()),
-                    Value::Number(n) => hasher.update(n.to_string().as_bytes()),
-                    Value::Bool(b) => hasher.update(if *b { "1" } else { "0" }.as_bytes()),
-                    other => hasher.update(other.to_string().as_bytes()),
-                }
-            }
-            hasher.update(b"|");
-        }
-        hex::encode(hasher.finalize())
-    }
-
     pub async fn get_row_metadata_since(
         &self,
         table: &str,
@@ -167,7 +83,7 @@ impl LocalSqlDataSource {
             )));
         }
 
-        let pk_expr = Self::build_pk_text_expr(&info.primary_key);
+        let pk_expr = adapter_common::build_pk_text_expr(&info.primary_key);
         let column_order: Vec<String> = info.columns.iter().map(|c| c.name.clone()).collect();
         let sql = format!(
             "SELECT *, {} AS __pk FROM \"{}\" WHERE \"{}\" > ?",
@@ -175,9 +91,9 @@ impl LocalSqlDataSource {
         );
         let params = vec![Value::String(since_timestamp.to_string())];
         let result = self.run(&sql, &params).await?;
-        let maps = self.rows_to_maps(&result.columns, &result.rows);
+        let maps = adapter_common::rows_to_maps(&result.columns, &result.rows);
 
-        Ok(Self::row_maps_to_metadata(
+        Ok(adapter_common::row_maps_to_metadata(
             &maps,
             &column_order,
             timestamp_column,
@@ -195,37 +111,6 @@ impl LocalSqlDataSource {
             .unwrap()
             .insert(table.to_string(), info.clone());
         Ok(info)
-    }
-
-    fn generate_batch_sql(
-        table: &str,
-        columns: &[String],
-        rows: &[HashMap<String, Value>],
-    ) -> (String, Vec<Value>) {
-        let col_list = columns
-            .iter()
-            .map(|c| format!("\"{}\"", c))
-            .collect::<Vec<_>>()
-            .join(", ");
-
-        let row_placeholder = format!("({})", vec!["?"; columns.len()].join(", "));
-        let all_placeholders = vec![row_placeholder.as_str(); rows.len()].join(", ");
-
-        let sql = format!(
-            "INSERT OR REPLACE INTO \"{}\" ({}) VALUES {}",
-            table, col_list, all_placeholders
-        );
-
-        let params: Vec<Value> = rows
-            .iter()
-            .flat_map(|row| {
-                columns
-                    .iter()
-                    .map(|c| row.get(c).cloned().unwrap_or(Value::Null))
-            })
-            .collect();
-
-        (sql, params)
     }
 }
 
@@ -318,13 +203,13 @@ impl DataSource for LocalSqlDataSource {
             )));
         }
 
-        let pk_expr = Self::build_pk_text_expr(&info.primary_key);
+        let pk_expr = adapter_common::build_pk_text_expr(&info.primary_key);
         let column_order: Vec<String> = info.columns.iter().map(|c| c.name.clone()).collect();
         let sql = format!("SELECT *, {} AS __pk FROM \"{}\"", pk_expr, table);
         let result = self.run(&sql, &[]).await?;
-        let maps = self.rows_to_maps(&result.columns, &result.rows);
+        let maps = adapter_common::rows_to_maps(&result.columns, &result.rows);
 
-        Ok(Self::row_maps_to_metadata(
+        Ok(adapter_common::row_maps_to_metadata(
             &maps,
             &column_order,
             timestamp_column,
@@ -342,7 +227,7 @@ impl DataSource for LocalSqlDataSource {
         }
 
         let info = self.cached_table_info(table).await?;
-        let pk_expr = Self::build_pk_text_expr(&info.primary_key);
+        let pk_expr = adapter_common::build_pk_text_expr(&info.primary_key);
 
         let placeholders: Vec<String> = pk_values.iter().map(|_| "?".to_string()).collect();
         let params: Vec<Value> = pk_values.iter().map(|v| Value::String(v.clone())).collect();
@@ -354,7 +239,7 @@ impl DataSource for LocalSqlDataSource {
         );
 
         let result = self.run(&sql, &params).await?;
-        Ok(self.rows_to_maps(&result.columns, &result.rows))
+        Ok(adapter_common::rows_to_maps(&result.columns, &result.rows))
     }
 
     async fn upsert_rows(&self, table: &str, rows: &[HashMap<String, Value>]) -> Result<usize> {
@@ -362,7 +247,7 @@ impl DataSource for LocalSqlDataSource {
             return Ok(0);
         }
         let columns: Vec<String> = rows[0].keys().cloned().collect();
-        let (sql, params) = Self::generate_batch_sql(table, &columns, rows);
+        let (sql, params) = adapter_common::generate_batch_sql(table, &columns, rows);
         self.run(&sql, &params).await.map_err(|e| {
             SyncError::Remote(format!(
                 "batch upsert failed for table '{}' ({} rows): {}",
